moved list selector into it's own component. renamed statusselector.
